requirements.txt
# Core web framework and dependencies
Flask==2.2.2
Jinja2==3.1.2
Werkzeug==2.2.2
click==8.1.3

# Computer vision and image processing
opencv-contrib-python==4.12.0.88
Pillow==9.4.0
numpy>=2.0.0,<2.3.0

# Raspberry Pi and GPIO libraries
RPi.GPIO==0.7.1a4
gpiozero==2.0.1
lgpio==0.2.2.0
picamera2==0.3.30
sense-hat==2.6.0
smbus2==0.4.2
spidev==3.5

# Networking and HTTP
requests==2.28.1
urllib3==1.26.12

# Data formats and serialization
PyYAML==6.0.2

# Development and testing tools
pytest==8.4.2
pytest-timeout==2.4.0

# System interaction
psutil==5.9.4
pyserial==3.5

# GUI framework (if needed)
PyQt5-sip==12.11.1
#PyQt5==5.15.9


# Environment configuration
python-dotenv==0.21.0
